Spawn Points, updated scripts and Ai
Placed all the spawn points, currently just set the amount of waves to 2 and only spawns one type of enemy. There was a bug with the navmesh agent which made the enemy's movement laggy, it was choppy, which delayed this commit. That bug has been resolved and the prototype can be handed in whenever. Game play is pretty fun when I (Angelo E.) playtested it.