Prototype of the boss and enemy projectile
I got too excited and started a boss script. So far I'm making it spawn enemies endlessly in waves, and the game ends when the boss is dead, it only goes vulnerable when the mini wave is defeated. I also made a basic sphere projectile for the bullet which has a trail renderer, it doesnt do anything special yet and I need to implement the shooting state for the enemy ai